fix errors from linux converter function

+++ src/type/ft_typeD.c
@@ -41,7 +41,7 @@ static char	*ft_nbnegative(char *s, char *tmp, s_option opt, size_t size)
 	}
 	else
 	{
-		abso = (opt.width < 0) ? ft_strlen(s) : opt.width;
+		abso = (opt.width < 0) ? (int)ft_strlen(s) : opt.width;
 		zero = (opt.flag.zero) ? abso - ft_strlen(s): 0;
 		start = (opt.flag.minus || opt.width < 0) ? 0 : size - FT_MAX(ft_strlen(s), zero);
 		tmp[start++] = '-';
@@ -74,7 +74,7 @@ static char	*ft_zeropad(char *s, char *tmp, s_option opt, size_t size)
 	}
 	else
 	{
-		abso = (opt.width < 0) ? ft_strlen(s) : opt.width;
+		abso = (opt.width < 0) ? (int)ft_strlen(s) : opt.width;
 		zero = (opt.flag.zero) ? abso - ft_strlen(s): 0;
 		start = (opt.flag.minus || opt.width < 0) ? 0 : size - FT_MAX(ft_strlen(s), zero);
 		while (i++ < zero)
